refactor: pulizia codice, mypy in CI, split test_config, legacy cleanup
  • raw/run.py: estratte 8 funzioni helper in raw/_fetch_utils.py
    con dispatch table per _fetch_payload e _infer_from_url unificata
  • core/config.py: rimosso bridge _compat_* (4 funzioni → _model_dump)
  • mypy: aggiunto in CI, 87 errori → 0, config in pyproject.toml
  • test_config.py: splittato in 3 file tematici (1237 → 335+434+23 righe)
  • legacy_aliases: default portato a False (non scrive più profile.json duplicato)
  • duckdb_read.py: rimosso re-export resolve_clean_read_cfg
  • conftest.py: rimosse referenze morte a file inesistenti
  • feature-stability.md: scout_url aggiornato a "removed"
  • Vari fix mypi: Pydantic override, str|None arg-type, import mancanti
  • README.md: riscritto con audience table, quickstart, MCP server, FAQ

Breaking: nessuno — 576 test passano, mypy zero errori
