+++ .github/workflows/ci.yml
@@ -22,6 +22,8 @@ jobs:
         run: python -m pip install -e ".[dev]"
       - name: Ruff
         run: python -m ruff check .
+      - name: Mypy
+        run: python -m mypy toolkit/
       - name: Pytest
         run: python -m pytest -p pytest_cov -q -p no:cacheprovider --cov=toolkit --cov-report=term-missing --cov-report=xml --cov-fail-under=70
       - name: Upload Coverage Artifact
