+++ docs/feature-stability.md
@@ -18,7 +18,7 @@ Questa matrice serve a chiarire cosa il toolkit considera percorso canonico, cos
 | artifact policy `minimal|standard|debug` | supported / advanced | tuning operativo |
 | `legacy_aliases` | compatibility only | non promuovere nei repo nuovi |
 | config legacy | compatibility only | usare `--strict-config` nei repo nuovi |
-| `scout_url` | deprecated | sostituito da `inspect url` |
+| ~~`scout_url`~~ | removed | sostituito da `inspect url` |
 | `inspect url` | experimental | scouting rapido di un URL pubblico e generazione scaffold YAML |
 | `inspect probe` | experimental | probe schema+stats di un endpoint SPARQL pre-candidate |
