+++ pyproject.toml
@@ -42,12 +42,51 @@ parquet = [
 ]
 dev = [
   "build>=1.4.3",
+  "mypy>=1.20",
   "pytest>=8.0",
   "pytest-cov>=5.0.0",
   "pytest-xdist>=3.0",
   "ruff>=0.3.0",
   "twine>=5.0.0",
   "xlwt>=1.0.0",
+]
+
+[tool.mypy]
+python_version = "3.10"
+warn_unused_configs = true
+warn_redundant_casts = true
+warn_unused_ignores = true
+warn_no_return = true
+warn_return_any = false
+no_implicit_optional = true
+ignore_missing_imports = true
+disallow_any_unimported = false
+disallow_untyped_defs = false
+
+# Modules affected by the dual Pydantic/dict config pattern.
+# Suppress known error codes until the dual config is resolved (tracked in audit).
+# TODO: remove these overrides after propagating Pydantic models through validate pipeline.
+[[tool.mypy.overrides]]
+module = [
+  "toolkit.mart.validate",
+  "toolkit.clean.validate",
+  "toolkit.cross.validate",
+  "toolkit.clean._column_rules",
+  "toolkit.core.layer_profile",
+]
+disable_error_code = [
+  "union-attr",
+  "index",
+  "arg-type",
+  "assignment",
+]
+
+[[tool.mypy.overrides]]
+module = [
+  "toolkit.cli.cmd_status",
+]
+disable_error_code = [
+  "attr-defined",
 ]
 
 [project.scripts]
